Revert "This closes #1541": Move PerKeyCombineFnRunner to runners-core
The following merge is reverted:

a52600b This closes #1541

The content of the merge was:

267edaa Move PerKeyCombineFnRunner to runners-core
d323d5d Update Dataflow worker to beam-master-20161212

Outstanding bugs in the new Dataflow worker were caught in postcommit.
