+++ runners/google-cloud-dataflow-java/src/main/resources/org/apache/beam/runners/dataflow/dataflow.properties
@@ -18,6 +18,6 @@
 
 environment.major.version=6
 
-worker.image.batch=dataflow.gcr.io/v1beta3/beam-java-batch:beam-master-20161212
+worker.image.batch=dataflow.gcr.io/v1beta3/beam-java-batch:beam-master-20161205
 
-worker.image.streaming=dataflow.gcr.io/v1beta3/beam-java-streaming:beam-master-20161212
+worker.image.streaming=dataflow.gcr.io/v1beta3/beam-java-streaming:beam-master-20161205
